[setupd] Fix the way of managing update process

+++ components/feral-setupd/scripts/fresh-debug.sh
@@ -1,5 +1,9 @@
 rm -f ~/.state/setupd
 cargo build
+if [ $? -ne 0 ]; then
+    echo "cargo build failed, stopping script"
+    exit 1
+fi
 sudo systemctl stop feral-setupd.service
 sudo cp ~/src/components/feral-setupd/target/debug/feral-setupd /usr/bin/feral-setupd
 sudo systemctl daemon-reload
