+++ components/feral-setupd/scripts/resume-debug.sh
@@ -1,4 +1,8 @@
 cargo build
+if [ $? -ne 0 ]; then
+    echo "cargo build failed, stopping script"
+    exit 1
+fi
 sudo systemctl stop feral-setupd.service
 sudo cp ~/feral-setupd/target/debug/feral-setupd /usr/bin/feral-setupd
 sudo systemctl daemon-reload
